refactor(hero-video): move HeroVideo out of clientComponents/
HeroVideo has no 'use client' directive and contains an async inner
component — it's a server component that was misfiled under
clientComponents/. Move to src/components/ to match the convention
documented in CLAUDE.md.
